+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add lint
 - Add customazing controls button prev/next
+- Rebuild table on change visibleRange
+- Update package dart_date
 
 # 0.0.10